Dual view code div now scrolls correctly in Firefox. Issue #4 from the tracker identified an issue occurring in Firefox, where if the dual view mode is selected (Click on "peek at code" when the blocks tab is active). This update fixes that issue, tested in Firefox 30.0, Chrome 35 and IE 11.
In addition the css file for the BlocklyDuino web app has been moved (and updated) from the media folder to the apps/blockyduino folder. This is because the media is a shared folder for different Blockly apps, and this specific css file is only relevant for the code webapp.

Fixes #4